Fix typing issues detected by CI

+++ modules/translation/i18n/i18n.go
@@ -8,7 +8,7 @@ import (
 	"io"
 )
 
-type PluralFormRule func(int) int
+type PluralFormRule func(int64) int
 
 const (
 	PluralFormZero = iota
